curso02/pyproject.toml
[tool.poetry]
name = "project2"
version = "0.1.0"
description = ""
authors = ["Wesley <wsdev200@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-decouple = "^3.8"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
pytest = "^7.2.2"
alembic = "^1.10.3"
psycopg2-binary = "^2.9.6"
httpx = "^0.23.3"
passlib = "^1.7.4"
python-jose = "^3.3.0"
python-multipart = "^0.0.6"
fastapi-pagination = "^0.12.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
